XCUITest formats 1 and 3 ares supported by xcuitest command
NEEDS TESTS
